refactor: move instance status client to a dedicated file
