+++ controllers/backup_controller.go
@@ -210,15 +210,19 @@ func (r *BackupReconciler) getBackupTargetPod(ctx context.Context, cluster apiv1
 				"pod", item.Pod.Name)
 			continue
 		}
-		if item.IsPrimary && cluster.Spec.Backup.Target != apiv1.BackupTargetStandby {
-			contextLogger.Debug("Primary Instance is elected as backup target",
-				"instance", item.Pod.Name)
-			return &item.Pod, nil
-		}
-		if !item.IsPrimary && cluster.Spec.Backup.Target == apiv1.BackupTargetStandby {
-			contextLogger.Debug("Standby Instance is elected as backup target",
-				"instance", item.Pod.Name)
-			return &item.Pod, nil
+		switch cluster.Spec.Backup.Target {
+		case apiv1.BackupTargetPrimary, "":
+			if item.IsPrimary {
+				contextLogger.Debug("Primary Instance is elected as backup target",
+					"instance", item.Pod.Name)
+				return &item.Pod, nil
+			}
+		case apiv1.BackupTargetStandby:
+			if !item.IsPrimary {
+				contextLogger.Debug("Standby Instance is elected as backup target",
+					"instance", item.Pod.Name)
+				return &item.Pod, nil
+			}
 		}
 	}
